+++ lib/airbrake-ruby/filter_chain.rb
@@ -8,22 +8,9 @@ module Airbrake
   # @since v1.0.0
   class FilterChain
     ##
-    # Replaces paths to gems with a placeholder.
-    # @return [Proc]
-    GEM_ROOT_FILTER = proc do |notice|
-      return unless defined?(Gem)
-
-      notice[:errors].each do |error|
-        Gem.path.each do |gem_path|
-          error[:backtrace].each do |frame|
-            # If the frame is unparseable, then 'file' is nil, thus nothing to
-            # filter (all frame's data is in 'function' instead).
-            next unless (file = frame[:file])
-            file.sub!(/\A#{gem_path}/, '[GEM_ROOT]'.freeze)
-          end
-        end
-      end
-    end
+    # @return [String] the namespace for filters, which are executed first,
+    #   before any other filters
+    LIB_NAMESPACE = '#<Airbrake::'.freeze
 
     ##
     # Filters to be executed last. By this time all permutations on a notice
@@ -35,34 +22,33 @@ module Airbrake
     ].freeze
 
     ##
-    # Skip over SystemExit exceptions, because they're just noise.
-    # @return [Proc]
-    SYSTEM_EXIT_FILTER = proc do |notice|
-      if notice[:errors].any? { |error| error[:type] == 'SystemExit' }
-        notice.ignore!
-      end
-    end
-
-    ##
     # @param [Airbrake::Config] config
     def initialize(config)
       @filters = []
       @keys_filters = []
 
-      [SYSTEM_EXIT_FILTER, GEM_ROOT_FILTER].each do |filter|
-        add_filter(filter)
+      [Airbrake::Filters::SystemExitFilter,
+       Airbrake::Filters::GemRootFilter].each do |filter|
+        @filters << filter.new
       end
 
       root_directory = config.root_directory
-      add_filter(root_directory_filter(root_directory)) if root_directory
+      return unless root_directory
+
+      @filters << Airbrake::Filters::RootDirectoryFilter.new(root_directory)
     end
 
     ##
     # Adds a filter to the filter chain.
+    #
     # @param [#call] filter The filter object (proc, class, module, etc)
+    # @return [void]
     def add_filter(filter)
       return @keys_filters << filter if KEYS_FILTERS.include?(filter.class)
-      @filters << filter
+      return @filters << filter unless filter.to_s.start_with?(LIB_NAMESPACE)
+
+      i = @filters.rindex { |f| f.to_s.start_with?(LIB_NAMESPACE) }
+      @filters.insert(i + 1, filter) if i
     end
 
     ##
@@ -75,19 +61,6 @@ module Airbrake
       (@filters + @keys_filters).each do |filter|
         break if notice.ignored?
         filter.call(notice)
-      end
-    end
-
-    private
-
-    def root_directory_filter(root_directory)
-      proc do |notice|
-        notice[:errors].each do |error|
-          error[:backtrace].each do |frame|
-            next unless (file = frame[:file])
-            file.sub!(/\A#{root_directory}/, '[PROJECT_ROOT]'.freeze)
-          end
-        end
       end
     end
   end